ref: remove bootstrap function
When the shutdown command is sent, the Sisyphus struct is dropped, which resulted in a hang/panic in case the shutdown command was sent while the task was bootstraping. By removing that function (and forcing the developer to put te functionality inside the spawn function), we don't have that undefined behavior.
Also:
- remove setting the  TaskStatus to Stopped when shutting down, as the SIsyphus struct has been stopped, thus the rx part of the watch channel as well, thus it's not possible to set the status any more